requirements_gemini.txt
# Requirements for Gemini API Integration
# Install with: pip install -r requirements_gemini.txt

# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Gemini API
google-generativeai>=0.3.0

# Additional utilities
pathlib2>=2.3.0
json5>=0.9.0

# Optional: for better error handling
requests>=2.28.0
urllib3>=1.26.0
